Simplify SPStore interface for the library user.
I am a little embarrassed by the previous state.  The library
interface exposed a tricky part of the application logic and left it
to the library users.  Now they only need to write, read, and delete
data, and when what should happen is left to the library.

This fixes at least two bugs: (1) `simpleCheckAgainstRequest` was
called twice inside `judge`, but at the time of the second call, the
first had already deleted the entry; (2) in module `Example`, the
entries were not deleted at all.